+++ src/lib/token.h
@@ -209,6 +209,7 @@ void token_unlock(token *t);
 CK_RV token_load_object(token *tok, CK_OBJECT_HANDLE key, tobject **loaded_tobj);
 
 CK_RV token_min_init(token *t);
+void token_reset(token *t);
 
 CK_RV token_init(token *t, CK_BYTE_PTR pin, CK_ULONG pin_len, CK_BYTE_PTR label);